% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDataPartition.R
\name{createDataPartition}
\alias{createDataPartition}
\title{creating  data partitioning}
\usage{
createDataPartition(y, p, list = FALSE)
}
\arguments{
\item{y}{a number}

\item{p}{a number}

\item{list}{a vector}
}
\value{
a number
}
\description{
creating  data partitioning
}
