% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData}
\alias{splitData}
\title{Splitting of data}
\usage{
splitData(data, indices)
}
\arguments{
\item{data}{a number}

\item{indices}{a number}
}
\value{
a number
}
\description{
Splitting of data
}
